Upgrade 002

+++ src/com/AlbertAbuav/facades/AdminFacade.java
@@ -40,10 +40,10 @@ public class AdminFacade extends ClientFacade {
     public void addCompany(Company company) throws invalidAdminException {
         if (Objects.isNull(company)) {
             throw new invalidAdminException("There is no customer like the one you entered!");
-        } else if (companiesDAO.isCompanyExistByEmail(company.getEmail())) {
-            throw new invalidAdminException("The email of the company you are trying to add already appears in the system.\nCompanies with the same email cannot be added!");
         } else if (companiesDAO.isCompanyExistByName(company.getName())) {
             throw new invalidAdminException("The name of the company you are trying to add already appears in the system.\nCompanies with the same name cannot be added!");
+        } else if (companiesDAO.isCompanyExistByEmail(company.getEmail())) {
+            throw new invalidAdminException("The email of the company you are trying to add already appears in the system.\nCompanies with the same email cannot be added!");
         }
         companiesDAO.addCompany(company);
     }
@@ -86,10 +86,12 @@ public class AdminFacade extends ClientFacade {
         if (companyCoupons.size() != 0) {
             for (Coupon coupon : companyCoupons) {
                 List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCouponsByCouponId(coupon.getId());
-                for (CustomersVsCoupons purchase : purchases) {
-                    couponsDAO.deleteCouponPurchase(purchase.getCustomerID(), purchase.getCouponID());
+                if (purchases.size() != 0) {
+                    for (CustomersVsCoupons purchase : purchases) {
+                        couponsDAO.deleteCouponPurchase(purchase.getCustomerID(), purchase.getCouponID());
+                    }
+                    couponsDAO.deleteCoupon(coupon.getId());
                 }
-                couponsDAO.deleteCoupon(coupon.getId());
             }
         }
         companiesDAO.deleteCompany(company.getId());
@@ -119,20 +121,12 @@ public class AdminFacade extends ClientFacade {
             throw new invalidAdminException("There is no id like you enter !");
         }
         Company company = companiesDAO.getSingleCompany(id);
-        if (company == null) {
+        if (Objects.isNull(company)) {
             throw new invalidAdminException("No Company was found by this id!");
         }
-        List<Coupon> dbCoupons = couponsDAO.getAllCoupons();
-        if (dbCoupons.size() != 0) {
-            List<Coupon> companyCoupons = new ArrayList<>();
-            for (Coupon dbCoupon : dbCoupons) {
-                if (dbCoupon.getCompanyID() == id) {
-                    companyCoupons.add(dbCoupon);
-                }
-            }
-            if (companyCoupons.size() != 0) {
-                company.setCoupons(companyCoupons);
-            }
+        List<Coupon> companyCoupons = couponsDAO.getAllCouponsOfSingleCompany(id);
+        if (companyCoupons.size() != 0) {
+            company.setCoupons(companyCoupons);
         }
         return company;
     }
@@ -144,10 +138,10 @@ public class AdminFacade extends ClientFacade {
      * @param customer Customer
      */
     public void addCustomer(Customer customer) throws invalidAdminException {
-        if (customer == null) {
+        if (Objects.isNull(customer)) {
             throw new invalidAdminException("There is no customer like you entered");
         }
-        if (customersDAO.getSingleCustomerByEmail(customer.getEmail())) {
+        if (customersDAO.isCustomerExistsByEmail(customer.getEmail())) {
             throw new invalidAdminException("The email of the customer you are trying to add already appears in the system.\nCustomers with the same email cannot be added.");
         }
         customersDAO.addCustomer(customer);
@@ -160,27 +154,17 @@ public class AdminFacade extends ClientFacade {
      * @param customer Customer
      */
     public void updateCustomer(Customer customer) throws invalidAdminException {
-        if (customersDAO.getAllCustomers() == null) {
-            throw new invalidAdminException("There are no customers in the system!");
-        }
-        if (customer == null) {
+        if (Objects.isNull(customer)) {
             throw new invalidAdminException("There is no customer like you entered");
         }
-        List<Customer> customerList = customersDAO.getAllCustomers();
-        Customer toCompare = null;
-        for (Customer customer1 : customerList) {
-            if (customer.getEmail().equals(customer1.getEmail())) {
-                toCompare = customer1;
+        if (!customersDAO.isCustomerExistsByEmail(customer.getEmail())) {
+            if (!customersDAO.isCustomerExistsById(customer.getId())) {
+                throw new invalidAdminException("There is no customer like you entered in the system or the email is incorrect!");
             }
+            customersDAO.updateCustomer(customer);
         }
-        if (toCompare == null) {
-            for (Customer customer1 : customerList) {
-                if (customer.getId() == customer1.getId()) {
-                    customersDAO.updateCustomer(customer);
-                    return;
-                }
-            }
-        } else if (customer.getId() != toCompare.getId()) {
+        Customer toCompare = customersDAO.getSingleCustomerByEmailB(customer.getEmail());
+        if (customer.getId() != toCompare.getId()) {
             throw new invalidAdminException("The customer id cannot be updated");
         }
         customersDAO.updateCustomer(customer);
@@ -193,20 +177,19 @@ public class AdminFacade extends ClientFacade {
      * @param customer Customer
      */
     public void deleteCustomer(Customer customer) throws invalidAdminException {
-        if (customer == null) {
+        if (Objects.isNull(customer)) {
             throw new invalidAdminException("This customer doesn't exists!");
         }
-        if (customersDAO.isCustomerExists(customer.getEmail(), customer.getPassword())) {
-            if (couponsDAO.getAllCustomersCoupons(customer.getId()) != null) {
-                List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(customer.getId());
-                for (CustomersVsCoupons purchase : purchases) {
-                    couponsDAO.deleteCouponPurchase(purchase.getCustomerID(), purchase.getCouponID());
-                }
-                customersDAO.deleteCustomer(customer.getId());
-            }
-        } else {
+        if (!customersDAO.isCustomerExists(customer.getEmail(), customer.getPassword())) {
             throw new invalidAdminException("There is no customer by the name \"" + customer.getFirstName() + " " + customer.getLastName() + "\" in the system!");
         }
+        List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(customer.getId());
+        if (purchases.size() != 0) {
+            for (CustomersVsCoupons purchase : purchases) {
+                couponsDAO.deleteCouponPurchase(purchase.getCustomerID(), purchase.getCouponID());
+            }
+        }
+        customersDAO.deleteCustomer(customer.getId());
     }
 
     /**
@@ -229,33 +212,38 @@ public class AdminFacade extends ClientFacade {
      * @return Customer
      */
     public Customer getSingleCustomer(int id) throws invalidAdminException {
-        if (customersDAO.getSingleCustomer(id) == null) {
-            throw new invalidAdminException("There is no customer with the id: " + id);
+        if (id <= 0) {
+            throw new invalidAdminException("There is no id like you enter !");
         }
         Customer customer = customersDAO.getSingleCustomer(id);
-        List<Coupon> customerCoupon = new ArrayList<>();
-        if (couponsDAO.getAllCustomersCoupons(id) != null) {
-            List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(id);
-            for (CustomersVsCoupons purchase : purchases) {
-                customerCoupon.add(couponsDAO.getSingleCoupon(purchase.getCouponID()));
-            }
-            if (customerCoupon.size() == 0) {
-                return customer;
-            }
-            customer.setCoupons(customerCoupon);
+        if (Objects.isNull(customer)) {
+            throw new invalidAdminException("There is no customer with the id: " + id);
         }
+        List<CustomersVsCoupons> purchases = couponsDAO.getAllCustomersCoupons(id);
+        if (purchases.size() == 0) {
+            return customer;
+        }
+        List<Coupon> customerCoupon = new ArrayList<>();
+        for (CustomersVsCoupons purchase : purchases) {
+            customerCoupon.add(couponsDAO.getSingleCoupon(purchase.getCouponID()));
+        }
+        if (customerCoupon.size() == 0) {
+            return customer;
+        }
+        customer.setCoupons(customerCoupon);
         return customer;
     }
 
     public List<CustomersVsCoupons> getAllCustomersVsCoupons(int customerID) throws invalidAdminException {
-        if (customersDAO.getSingleCustomer(customerID) == null) {
+        if (customerID <= 0) {
+            throw new invalidAdminException("There is no id like you enter !");
+        }
+        if (!customersDAO.isCustomerExistsById(customerID)) {
             throw new invalidAdminException("There is no customer with the id: " + customerID);
         }
-        List<CustomersVsCoupons> purchase = null;
-        if (couponsDAO.getAllCustomersCoupons(customerID) != null) {
-            purchase = couponsDAO.getAllCustomersCoupons(customerID);
-        } else {
-            throw new invalidAdminException("There are no customer purchases!");
+        List<CustomersVsCoupons> purchase = couponsDAO.getAllCustomersCoupons(customerID);
+        if (purchase.size() == 0) {
+            throw new invalidAdminException("There are no customer purchases to customer:\n" + customersDAO.getSingleCustomer(customerID));
         }
         return purchase;
     }
